pyproject.toml
[project]
name = "rockgarden"
version = "0.4.11"
description = "Obsidian-compatible static site generator"
readme = "README.md"
authors = [
    { name = "Dave Forgac", email = "tylerdave@tylerdave.com" }
]
requires-python = ">=3.13"
license = "MIT"
keywords = ["obsidian", "static-site-generator", "markdown"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "typer>=0.9.0",
    "jinja2>=3.1.0",
    "markdown-it-py>=3.0.0",
    "python-frontmatter>=1.0.0",
    "linkify-it-py>=2.0.3",
    "tzdata>=2024.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
]

[project.scripts]
rockgarden = "rockgarden.cli:main"
rgdn = "rockgarden.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rockgarden"]

[dependency-groups]
dev = [
    "commitizen>=4.11.6",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::DeprecationWarning:frontmatter"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.4.11"
version_files = [
    "pyproject.toml:^version",
    "src/rockgarden/__init__.py:^__version__",
]
tag_format = "v$version"
pre_bump_hooks = [
    "uv lock",
    "git add uv.lock",
]
